+++ lib/aegis/client.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require 'aws-sdk-athena'
-
 module Aegis
   class Client
     QUERY_STATUS_MAPPING = {
@@ -12,16 +10,14 @@ module Aegis
       'CANCELLED' => Aegis::QueryStatus::CANCELLED
     }.freeze
 
-    S3_URL_PATTERN = %r{^s3://(?<bucket>\S+?)/(?<key>\S+)$}.freeze
-
     DEFAULT_QUERY_STATUS_BACKOFF = ->(attempt) { 1.5**attempt - 1 }
 
-    private_constant :QUERY_STATUS_MAPPING, :S3_URL_PATTERN, :DEFAULT_QUERY_STATUS_BACKOFF
+    private_constant :QUERY_STATUS_MAPPING, :DEFAULT_QUERY_STATUS_BACKOFF
 
-    def initialize(aws_athena_client: nil, configuration: Aegis.configuration)
-      @configuration = configuration
-      @aws_athena_client = aws_athena_client || Aws::Athena::Client.new(default_athena_client_config(configuration))
-      @query_status_backoff = configuration.query_status_backoff || DEFAULT_QUERY_STATUS_BACKOFF
+    def initialize(aws_client_provider: Aegis::AwsClientProvider.new, s3_location_parser: Aegis::S3LocationParser.new)
+      @aws_athena_client = aws_client_provider.athena_client
+      @s3_location_parser = s3_location_parser
+      @query_status_backoff = Aegis.configuration.query_status_backoff || DEFAULT_QUERY_STATUS_BACKOFF
     end
 
     def database(database_name)
@@ -33,7 +29,7 @@ module Aegis
         query_execution_params(query, work_group, database, output_location)
       ).query_execution_id
 
-      return query_execution_id if async
+      return query_status(query_execution_id) if Aegis.mode.async(async)
 
       query_status = wait_for_query_to_finish(query_execution_id)
 
@@ -44,32 +40,28 @@ module Aegis
 
     def query_status(query_execution_id)
       resp = aws_athena_client.get_query_execution({query_execution_id: query_execution_id})
+
+      query_execution = resp.query_execution
+
       Aegis::QueryStatus.new(
-        QUERY_STATUS_MAPPING.fetch(resp.query_execution.status.state),
-        resp.query_execution.status.state_change_reason,
-        parse_output_location(resp)
+        query_execution.query_execution_id,
+        QUERY_STATUS_MAPPING.fetch(query_execution.status.state),
+        query_execution.status.state_change_reason,
+        parse_output_location(query_execution)
       )
     end
 
     private
 
-    attr_reader :aws_athena_client, :configuration, :query_status_backoff
-
-    def default_athena_client_config(configuration)
-      config = {}
-      config[:region] = configuration.aws_region if configuration.aws_region
-      config[:access_key_id] = configuration.aws_access_key_id if configuration.aws_access_key_id
-      config[:secret_access_key] = configuration.aws_secret_access_key if configuration.aws_secret_access_key
-      config
-    end
+    attr_reader :aws_athena_client, :s3_location_parser, :query_status_backoff
 
     def query_execution_params(query, work_group, database, output_location)
-      work_group_params = work_group || configuration.work_group
+      work_group_params = work_group || Aegis.configuration.work_group
 
       params = {query_string: query}
       params[:work_group] = work_group_params if work_group_params
-      params[:query_execution_context] = {database: database} if database
-      params[:result_configuration] = {output_location: output_location} if output_location
+      params[:query_execution_context] = {database: database_name(database)} if database
+      params[:result_configuration] = {output_location: translate_path(output_location)} if output_location
       params
     end
 
@@ -84,12 +76,20 @@ module Aegis
       end
     end
 
-    def parse_output_location(resp)
-      url = resp.query_execution.result_configuration.output_location
+    def parse_output_location(query_execution)
+      url = query_execution.result_configuration.output_location
 
-      matched_data = S3_URL_PATTERN.match(url)
+      bucket, path = s3_location_parser.parse_url(url)
 
-      QueryOutputLocation.new(url, matched_data[:bucket], matched_data[:key])
+      QueryOutputLocation.new(url, bucket, path)
+    end
+
+    def translate_path(s3_url)
+      Aegis.mode.s3_path(s3_url)
+    end
+
+    def database_name(name)
+      Aegis.mode.database_name(name)
     end
   end
 end