+++ lib/aegis/types/integer_serializer.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Aegis
+  module Types
+    class IntegerSerializer
+      def literal(integer)
+        integer.to_s
+      end
+
+      def dump(integer)
+        integer.to_s
+      end
+
+      def load(string)
+        string.to_i
+      end
+    end
+  end
+end