+++ lib/aegis/types/string_serializer.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Aegis
+  module Types
+    class StringSerializer
+      def literal(string)
+        "'#{string.gsub("'", "''")}'"
+      end
+
+      def dump(string)
+        string
+      end
+
+      def load(string)
+        string
+      end
+    end
+  end
+end