+++ spec/unit/aegis/query_status_spec.rb
@@ -3,7 +3,8 @@
 require 'spec_helper'
 
 RSpec.describe Aegis::QueryStatus do
-  let(:query_status) { described_class.new(status, message, output_location) }
+  let(:query_status) { described_class.new(id, status, message, output_location) }
+  let(:id) { '123' }
   let(:message) { nil }
   let(:output_location) { 's3://bucket/location' }
 
@@ -76,6 +77,28 @@ RSpec.describe Aegis::QueryStatus do
 
     context 'when status running' do
       let(:status) { :running }
+
+      it { is_expected.to be(true) }
+    end
+
+    context 'when status finished' do
+      let(:status) { :finished }
+
+      it { is_expected.to be(false) }
+    end
+  end
+
+  describe '#in_progress?' do
+    subject { query_status.in_progress? }
+
+    context 'when status running' do
+      let(:status) { :running }
+
+      it { is_expected.to be(true) }
+    end
+
+    context 'when status queued' do
+      let(:status) { :queued }
 
       it { is_expected.to be(true) }
     end